+++ sep490-enterprise/build.gradle
@@ -25,6 +25,7 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-security'
 	implementation 'org.springframework.boot:spring-boot-starter-web'
 	implementation 'org.springframework.kafka:spring-kafka'
+	implementation 'vn.payos:payos-java:1.0.3'
 
 	// Testing
 	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
