+++ sep490-enterprise/src/main/java/enterprise/dtos/CreditPurchaseDTO.java
@@ -1,0 +1,4 @@
+package enterprise.dtos;
+
+public record CreditPurchaseDTO(Long code, Integer amount, Integer price) {
+}
