+++ sep490-enterprise/src/main/resources/application.yml
@@ -58,3 +58,11 @@ logging:
         type:
           descriptor:
             sql: ${HIBERNATE_SQL_PARAM_LOG_LEVEL:TRACE}
+
+payment:
+  payos:
+    client_id: ${PAYOS_CLIENT_ID:not_a_real_key_to_prevent_null}
+    api_key: ${PAYOS_API_KEY:not_a_real_key_to_prevent_null}
+    checksum_key: ${PAYOS_CHECKSUM_KEY:not_a_real_key_to_prevent_null}
+    returnPath: /success-payment
+    cancelPath: /fail-payment
