--- sep490-idp/src/main/java/green_buildings/idp/service/PaymentService.java
@@ -1,9 +1,0 @@
-package green_buildings.idp.service;
-
-import green_buildings.idp.dto.CreditPurchaseItem;
-import vn.payos.type.CheckoutResponseData;
-
-public interface PaymentService {
-    public CheckoutResponseData getCheckoutData(CreditPurchaseItem creditPurchaseItem);
-    
-}
